Rename files based on new library name
