+++ covjsonkit/CoverageCollection.py
