+++ covjsonkit/__init__.py
@@ -1,0 +1,5 @@
+import covjsonkit.api
+import covjsonkit.decoder.TimeSeries
+import covjsonkit.decoder.VerticalProfile
+import covjsonkit.encoder.TimeSeries
+import covjsonkit.encoder.VerticalProfile
