+++ covjsonkit/decoder/BoundingBox.py
