+++ covjsonkit/decoder/Frame.py
