+++ covjsonkit/decoder/Path.py
