+++ covjsonkit/decoder/Shapefile.py
