+++ covjsonkit/decoder/TimeSeries.py
