+++ covjsonkit/encoder/BoundingBox.py
