+++ covjsonkit/encoder/Frame.py
