+++ covjsonkit/encoder/Path.py
