+++ covjsonkit/encoder/Shapefile.py
