+++ covjsonkit/encoder/TimeSeries.py
