+++ covjsonkit/encoder/Wkt.py
