--- eccovjson/__init__.py
@@ -1,5 +1,0 @@
-import eccovjson.api
-import eccovjson.decoder.TimeSeries
-import eccovjson.decoder.VerticalProfile
-import eccovjson.encoder.TimeSeries
-import eccovjson.encoder.VerticalProfile
