+++ tests/test_decoder_vertical_profile.py
@@ -1,4 +1,4 @@
-from eccovjson.api import Eccovjson
+from covjsonkit.api import Eccovjson
 
 
 class TestDecoder:
