+++ tests/test_encoder_frame.py
@@ -1,7 +1,7 @@
 from covjson_pydantic.coverage import CoverageCollection
 from covjson_pydantic.domain import DomainType
 
-from eccovjson.api import Eccovjson
+from covjsonkit.api import Eccovjson
 
 
 def get_timestamps(start_dt, end_dt, delta):
