+++ tests/test_encoder_path.py
@@ -2,7 +2,7 @@ import pytest
 from covjson_pydantic.coverage import CoverageCollection
 from covjson_pydantic.domain import DomainType
 
-from eccovjson.api import Eccovjson
+from covjsonkit.api import Eccovjson
 
 
 def get_timestamps(start_dt, end_dt, delta):
