+++ tests/test_encoder_time_series.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta
 from covjson_pydantic.coverage import CoverageCollection
 from covjson_pydantic.domain import DomainType
 
-from eccovjson.api import Eccovjson
+from covjsonkit.api import Eccovjson
 
 
 def get_timestamps(start_dt, end_dt, delta):
